requirements.txt
paho-mqtt==1.1
psutil==6.0.0
opencv-python==4.10.0.84